requirements.txt
# Core Dependencies - AI Voice Agent for Lead Generation
streamlit==1.50.0
google-generativeai==0.8.5
python-dotenv==1.1.1
supabase==2.20.0

# Audio Processing
SpeechRecognition==3.14.3
pyttsx3==2.99
pydub==0.25.1